fix: make hero section auth-aware on discover page
Converts page.tsx to a server component that reads the session via auth(),
then passes isAuthenticated to a new HomePageClient component. Signed-in
users now see a link to their saved shops/alerts instead of a sign-in prompt.

Closes #152
